Make createLocation an instance method
A single, global createLocation method cannot be aware of history
enhancers like useQueries. So users need to manually modify location
objects when creating them programmatically instead of getting them
in a listen callback.

This commit makes createLocation an instance method of history objects
and deprecates the createLocation method on the top-level exports.