Add rawPathname to location
This property is set prior to decoding the pathname. For browser and memory histories, the rawPathname will be the encoded version of the provided pathname (generated using an anchor element). For memory histories, the rawPathname is the same as the pathname.

When creating a URL (using createPath), the rawPathname property will be preferred over the pathname property.