Pass the raw, not decoded, path to browser
The path is first decoded in createLocation before being passed to the
browser through history.pushState or location.href to be encoded again.
This does not work for percent signs, so the original, presumably
encoded, path is used instead.

An alternative solution would be to add a configurable encode function
to be called in createPath.