set initial memory history key to "default"
Documentation states that the initial entry always has key `"default"`. However on memory history this is not the case.

We check that the key is `"default"` to set a default behavior when the user clicks a "back" button to go back to home, but arrived at the page from a direct link. This works perfectly in the browser, but failed in the test because the key was not "default" on MemoryRouter.